perf(parser): add find cache
This is part of #3702 investigation and try to found cpu usage issue
